# exp/orderbook: Remove duffcopy'ies in `consumeOffers*` functions