<!-- If you're making a doc PR or something tiny where the below is irrelevant, delete this
template and use a short description, but in your description aim to include both what the
change is, and why it is being made, with enough context for anyone to understand. -->

<details>
  <summary>PR Checklist</summary>
  
### PR Structure

* [x] This PR has reasonably narrow scope (if not, break it down into smaller PRs).
* [x] This PR avoids mixing refactoring changes with feature changes (split into two PRs
  otherwise).
* [x] This PR's title starts with name of package that is most changed in the PR, ex.
  `services/friendbot`, or `all` or `doc` if the changes are broad or impact many
  packages.

### Thoroughness

* [ ] This PR adds tests for the most critical parts of the new functionality or fixes.
* [ ] I've updated any docs ([developer docs](https://www.stellar.org/developers/reference/), `.md`
  files, etc... affected by this change). Take a look in the `docs` folder for a given service,
  like [this one](https://github.com/stellar/go/tree/master/services/horizon/internal/docs).

### Release planning

* [ ] I've updated the relevant CHANGELOG ([here](services/horizon/CHANGELOG.md) for Horizon) if
  needed with deprecations, added features, breaking changes, and DB schema changes.
* [ ] I've decided if this PR requires a new major/minor version according to
  [semver](https://semver.org/), or if it's mainly a patch change. The PR is targeted at the next
  release branch if it's not a patch change.
</details>

### What

This commit removes `runtime.duffcopy` usage in `consumeOffersForBuyingAsset` and `consumeOffersForSellingAsset` increasing the execution speed.

### Why

When using `range` over slices Golang copies each value to a temporary object. This usually isn't a performance issue however `consumeOffersForSellingAsset` and `consumeOffersForBuyingAsset` are used in `dfs` that is recursively called thousands times when searching for paths, it takes a relatively big percentage of running time.

<img width="468" alt="Screenshot 2020-02-07 at 18 05 20" src="https://user-images.githubusercontent.com/464938/74049517-916dad80-49d4-11ea-8bc0-13afada81de9.png">

Before changes:
```
(pprof) tree github.com/stellar/go/exp/orderbook.consumeOffersForBuyingAsset
Active filters:
   focus=github.com/stellar/go/exp/orderbook.consumeOffersForBuyingAsset
Showing nodes accounting for 23.43s, 19.49% of 120.22s total
Dropped 1 node (cum <= 0.60s)
----------------------------------------------------------+-------------
      flat  flat%   sum%        cum   cum%   calls calls% + context 	 	 
----------------------------------------------------------+-------------
                                            12.06s   100% |   github.com/stellar/go/exp/orderbook.consumeOffersForBuyingAsset
    12.06s 10.03% 10.03%     12.06s 10.03%                | runtime.duffcopy
```
After changes:
```
(pprof) tree github.com/stellar/go/exp/orderbook.consumeOffersForBuyingAsset
Active filters:
   focus=github.com/stellar/go/exp/orderbook.consumeOffersForBuyingAsset
Showing nodes accounting for 18.34s, 15.63% of 117.37s total
Dropped 1 node (cum <= 0.59s)
----------------------------------------------------------+-------------
      flat  flat%   sum%        cum   cum%   calls calls% + context 	 	 
----------------------------------------------------------+-------------
                                            18.35s   100% |   github.com/stellar/go/exp/orderbook.(*buyingGraphSearchState).consumeOffers
    11.57s  9.86%  9.86%     18.35s 15.63%                | github.com/stellar/go/exp/orderbook.consumeOffersForBuyingAsset
                                             5.54s 30.19% |   github.com/stellar/go/price.MulFractionRoundDown
                                             1.24s  6.76% |   github.com/stellar/go/price.ConvertToBuyingUnits
```